pc-agent/requirements.txt
psutil==5.9.8
pywin32==307
requests==2.31.0
python-dotenv==1.0.1
